+++ src/main/java/com/chargebee/models/Transaction.java
@@ -16,6 +16,7 @@ public class Transaction extends Resource<Transaction> {
         CASH,
         CHECK,
         BANK_TRANSFER,
+        AMAZON_PAYMENTS,
         OTHER,
         _UNKNOWN; /*Indicates unexpected value for this enum. You can get this when there is a
         java-client version incompatibility. We suggest you to upgrade to the latest version */